staking refactoring, update state, add basic tokens support
