+++ src/components/composed/CurrentStaking/CurrentStaking.js
@@ -8,16 +8,21 @@ import { ML } from '@Helpers'
 import { LocalStorageService } from '@Storage'
 import { AppInfo } from '@Constants'
 
-import { TransactionContext, SettingsContext, AccountContext } from '@Contexts'
+import { SettingsContext } from '@Contexts'
 
 import './CurrentStaking.css'
-import Timer from '../../basic/Timer/Timer'
-import { useEffectOnce } from '../../../hooks/etc/useEffectOnce'
+import { useNavigate, useParams } from 'react-router-dom'
 
 const CurrentStaking = ({ addressList }) => {
+  const navigate = useNavigate()
+  const { coinType } = useParams()
   const { networkType } = useContext(SettingsContext)
-  const { setDelegationStep } = useContext(TransactionContext)
-  const { walletType } = useContext(AccountContext)
+
+  const walletType = {
+    name: coinType,
+    ticker: coinType === 'Mintlayer' ? 'ML' : 'BTC',
+    network: coinType === 'Bitcoin' ? 'bitcoin' : 'mintlayer',
+  }
 
   const account = LocalStorageService.getItem('unlockedAccount')
   const accountName = account.name
@@ -26,26 +31,13 @@ const CurrentStaking = ({ addressList }) => {
     LocalStorageService.getItem(unconfirmedTransactionString) &&
     walletType.name === 'Mintlayer'
 
-  const unconfirmedTransactions = LocalStorageService.getItem(
-    unconfirmedTransactionString,
-  )
-
-  const revalidate =
-    isUncofermedTransaction &&
-    unconfirmedTransactions.mode === AppInfo.ML_TRANSACTION_MODES.DELEGATION
   const {
     mlDelegationList,
     mlDelegationsBalance,
-    getDelegations,
-    getTransactions,
   } = useMlWalletInfo(addressList)
 
-  useEffectOnce(() => {
-    getDelegations()
-  })
-
-  const onNextButtonClick = () => {
-    setDelegationStep(2)
+  const onDelegationCreateButtonClick = () => {
+    navigate('/wallet/' + walletType.name + '/staking/create-delegation')
   }
   const stakingGuideLink =
     'https://mintlayer.info/en/Guides/Staking/browser-extension'
@@ -70,18 +62,6 @@ const CurrentStaking = ({ addressList }) => {
             Total staked: {ML.getAmountInCoins(mlDelegationsBalance)} ML
           </p>
         </div>
-        <div>
-          {revalidate && (
-            <Timer
-              onTimerEnd={() => {
-                getDelegations()
-                getTransactions()
-              }}
-              duration={10000}
-              repeat={revalidate}
-            />
-          )}
-        </div>
         <a
           href={poolListLink}
           target="_blank"
@@ -93,7 +73,7 @@ const CurrentStaking = ({ addressList }) => {
       <Wallet.DelegationList delegationsList={mlDelegationList} />
       <div className="delegation-button-wrapper">
         <Button
-          onClickHandle={onNextButtonClick}
+          onClickHandle={onDelegationCreateButtonClick}
           extraStyleClasses={['delegation-button']}
           disabled={isUncofermedTransaction}
         >
