+++ src/components/composed/Header/Header.js
@@ -8,12 +8,11 @@ import { ReactComponent as ExpandImg } from '@Assets/images/icon-expand.svg'
 import { ReactComponent as SettingsImg } from '@Assets/images/settings.svg'
 
 import { Button, Logo, Tooltip } from '@BasicComponents'
-import { AccountContext, NetworkContext } from '@Contexts'
+import { AccountContext } from '@Contexts'
 
 import './Header.css'
 
 const Header = ({ customBackAction, noBackButton = false }) => {
-  const network = useContext(NetworkContext)
   const [unlocked, setUnlocked] = useState(false)
   const [tooltipVisible, setTooltipVisible] = useState(false)
   const tooltipMessage = 'Expand view'
@@ -61,10 +60,6 @@ const Header = ({ customBackAction, noBackButton = false }) => {
 
   return (
     <header data-testid="header-container">
-      <pre style={{ position: 'absolute', top: 1, left: 1 }}>
-        {JSON.stringify(network)}
-      </pre>
-
       {!noBackButton && (
         <Button
           alternate
