+++ src/pages/CreateDelegation/CreateDelegation.css
@@ -1,0 +1,3 @@
+.page {
+  margin: 0;
+}
