+++ src/pages/DelegationStake/DelegationStake.css
@@ -1,0 +1,3 @@
+.page {
+  margin: 0;
+}
